+++ pylastfm/client.py
@@ -9,7 +9,7 @@ from pylastfm.response.common import PaginateMixin
 from pylastfm import auth, constants, error
 from pylastfm.util import (Signer, PaginatedIterator, nested_get, nested_in,
                            nested_set, ceildiv)
-from pylastfm.api import artist, user, track, auth as apiauth
+from pylastfm.api import artist, library, user, track, auth as apiauth
 
 
 def prefixed(prfx, *methods):
@@ -140,6 +140,7 @@ class LastFM(object):
 
         # Exposed API objects
         self.artist = artist.Artist(self)
+        self.library = library.Library(self)
         self.user = user.User(self)
         self.track = track.Track(self)
         self.auth = apiauth.Auth(self)