+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     - `user.get_shouts`
     - `user.shout`
 - Add artist API resource
+- Add library API resource
 
 0.1.1
 -----